Consolidated Relief Allowance (CRA)

consolidated_relief() computes the Consolidated Relief Allowance (CRA), a key tax relief under Nigeria’s Personal Income Tax Act (PITA) 2020.

Usage

consolidated_relief(gross_income, period = "yearly")

Arguments

gross_income

Numeric. The income on which CRA is computed, typically the result of gross_income().

period

Character. The period for which CRA is calculated. Accepts "yearly" (default) or "monthly".

Value

A numeric value representing the Consolidated Relief Allowance (CRA).

Details

The CRA is designed to reduce the taxable income of an individual, ensuring that a portion of their earnings is exempt from taxation before applying the income tax rates.

According to Section 33(2) of PITA 2020, the Consolidated Relief Allowance (CRA) is defined as:

"There shall be allowed a consolidated relief allowance of ₦200,000 subject to a minimum of 1% of gross income, whichever is higher, plus 20% of the gross income. The balance shall be taxable in accordance with the income table in the Sixth Schedule to this Act."

This means the CRA is computed using the formula:

$$ CRA = \begin{cases} \text{Gross Income} \times 20\% + 200,000, & \text{if } 1\% \times \text{Gross Income} \leq 200,000 \\ \text{Gross Income} \times 20\% + 1\% \times \text{Gross Income}, & \text{if } 1\% \times \text{Gross Income} > 200,000 \end{cases} $$

This logic can also be expressed in Excel as:

CRA = IF(gross_income * 1% > 200000, gross_income * 1% + gross_income * 20%, 200000 + gross_income * 20%)

In R, this is implemented using the ifelse() function.

See also

Examples

# Example 1: If 1% of gross income is greater than 200,000 Naira
consolidated_relief(5000000)
#> [1] 1200000
# Returns: (1% * 5000000) + (20% * 5000000) = 50,000 + 1,000,000 = 1,050,000

# Example 2: If 1% of gross income is less than or equal to 200,000 Naira
consolidated_relief(2000000)
#> [1] 6e+05
# Returns: 200,000 + (20% * 2000000) = 200,000 + 400,000 = 600,000

# Example 3: Compute monthly CRA for a gross income of 3,600,000 Naira
consolidated_relief(3600000, "monthly")
#> [1] 756000
# 1% of 3,600,000 = 36,000 (greater than 16,666.67, so we take 36,000)
# Monthly CRA = (36,000 + 20% * 3,600,000) = 36,000 + 720,000 = 756,000