Employee Tax Liability (PITA 2020)

The employee_tax_liability() function computes the Personal Income Tax (PAYE) liability for an individual based on Nigeria’s Personal Income Tax Act (PITA) 2020 graduated tax brackets. For more details, refer to the Finance Act 2020 document.

Usage

employee_tax_liability(taxable_income, period = "monthly")

Arguments

taxable_income

Numeric. The taxable income after applying all allowable deductions.

period

Character. Specifies whether the tax is calculated on a yearly or monthly basis. Default is "yearly".

Value

A numeric value representing the total PAYE tax liability.

Details

The tax calculation follows Nigeria's progressive tax system:

Income below 300,000 NGN is taxed at 1%.

The first 300,000 NGN is taxed at 7%.

The next 300,000 NGN is taxed at 11%.

The next 500,000 NGN is taxed at 15%.

The next 500,000 NGN is taxed at 19%.

The next 1,600,000 NGN is taxed at 21%.

Any income above 3,200,000 NGN is taxed at 24%.

Note: According to the Finance Act 2020, employees earning minimum wage are exempt from PAYE tax.

Examples

# Example: Employee with a taxable income of 4,500,000 NGN (Yearly)
employee_tax_liability(4500000)
#> [1] 872000

# Example: Employee with a taxable income of 375,000 NGN (Monthly)
employee_tax_liability(375000, period = "monthly")
#> [1] 72666.63