Payroll Dataset for FMC

fmc2017 contains payroll data for employees at FMC. It includes basic salary, allowances, and other financial details.

Usage

fmc2017

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1148 rows and 14 columns.

Source

FMC Payroll System

Details

@format A data frame with n rows and m columns:

S/N

Numeric. Serial number of employee record.

Name

Character. Employee name.

Staff_No

Numeric. Employee ID.

ConM/H

Numeric. The employee's grade level.

Basic_Salary

Numeric. The employee's basic salary.

Teaching_Allowance

Numeric. Allowance for teaching staff.

Salary_Arrears

Numeric. Salary arrears, if applicable.

other_allowances

Numeric. Additional allowances.

Overtime

Numeric. Overtime payment.

Shifting

Numeric. Shift-related payments.

Specialist

Numeric. Extra pay for specialist work.

Hazard_Allowance

Numeric. Extra pay for hazardous work.

Call_Allowance

Numeric. Allowance for call staff.

Grosspay

Numeric. Employee's net salary after deductions.

Examples

data(fmc2017)
head(fmc2017)
#> # A tibble: 6 × 14
#>   `S/N` Name    Staff_No `ConM/H` Basic_Salary Teaching_Allowance Salary_Arrears
#>   <dbl> <chr>   <chr>    <chr>           <dbl>              <dbl>          <dbl>
#> 1     1 DR. IN… "SS\\53… "MD"          120367.             37628.              0
#> 2     2 DR. WA… "SS\\23… "15\\9"       428250              37628.              0
#> 3     3 DR. EK… "SS\\19… "15\\9"       428250              37628.              0
#> 4     4 Dr. Ar… "SS\\94… "15\\9"       428250              37628.              0
#> 5     5 DR. MA… "SS\\23… "15\\9"       428250              37628.              0
#> 6     6 DR. ST… "SS\\26… "14\\9"       351483.             30502.              0
#> # ℹ 7 more variables: other_allowances <dbl>, Overtime <dbl>, Shifting <dbl>,
#> #   Specialist <dbl>, Hazard_Allowance <dbl>, Call_Allowance <dbl>,
#> #   GrossPay <dbl>